[FEATURE] Complete target-based find modules and CMake-based pyopenms builds

+++ src/pyOpenMS/pxds/AASequence.pxd
@@ -24,7 +24,7 @@ cdef extern from "<OpenMS/CHEMISTRY/AASequence.h>" namespace "OpenMS":
         AASequence iadd(AASequence) nogil except + # wrap-as:operator+=
 
         # Note that this is a const-ref, so we cannot easily set residues
-        Residue operator[](int) nogil except + # wrap-upper-limit:size()
+        Residue operator[](size_t) nogil except + # wrap-upper-limit:size()
 
         # check if sequence is empty
         bool empty() nogil except + # wrap-doc:Check if sequence is empty
