+++ src/pyOpenMS/pxds/IntegerDataArray.pxd
@@ -23,7 +23,7 @@ cdef extern from "<OpenMS/METADATA/DataArrays.h>" namespace "OpenMS::DataArrays"
         Size size() nogil except +
         void resize(size_t n) nogil except +
         void reserve(size_t n) nogil except +
-        Int& operator[](int) nogil except + # wrap-ignore
+        Int& operator[](size_t) nogil except + # wrap-ignore
         void clear() nogil except +
         void push_back(Int) nogil except +
